ai-agent-cs/requirements.txt
# Core Web & API
fastapi>=0.110.0
uvicorn>=0.27.1
python-dotenv
loguru
requests
httpx>=0.28.1
anyio

# AI & Database
google-genai>=1.0.0
faiss-cpu
numpy
nest-asyncio

# Telegram Bot
python-telegram-bot

# Utils
pydantic>=2.0.0
Jinja2
PyYAML
tqdm
certifi
charset-normalizer
idna
urllib3
